fix(morpho-sdk): restrict Midnight authorization grants
