+++ packages/morpho-sdk/src/actions/requirements/midnight/getMidnightAuthorizationRequirement.ts
@@ -1,22 +1,13 @@
 import { midnightAbi } from "@morpho-org/midnight-sdk";
-import {
-  deepFreeze,
-  getChainAddress,
-  getChainAddresses,
-} from "@morpho-org/morpho-ts";
-import {
-  type Address,
-  type Client,
-  encodeFunctionData,
-  isAddressEqual,
-} from "viem";
+import { deepFreeze, getChainAddress } from "@morpho-org/morpho-ts";
+import { type Address, type Client, encodeFunctionData } from "viem";
 import { readContract } from "viem/actions";
 import { validateChainId } from "../../../helpers/index.js";
+import { validateMidnightAuthorizationTarget } from "../../../helpers/validateMidnightAuthorizationTarget.js";
 import type {
   MidnightAuthorizationAction,
   Transaction,
 } from "../../../types/index.js";
-import { UnsupportedMidnightAuthorizationTargetError } from "../../../types/index.js";
 
 /** Parameters for {@link getMidnightAuthorizationRequirement}. */
 export interface GetMidnightAuthorizationRequirementParams {
@@ -64,25 +55,11 @@ export const getMidnightAuthorizationRequirement = async (
   params: GetMidnightAuthorizationRequirementParams,
 ): Promise<Readonly<Transaction<MidnightAuthorizationAction>> | null> => {
   validateChainId(params.viemClient.chain?.id, params.chainId);
-
-  const { midnightBundles, ecrecoverRatifier, setterRatifier } =
-    getChainAddresses(params.chainId);
-  const supportedTargets = [
-    midnightBundles,
-    ecrecoverRatifier,
-    setterRatifier,
-  ].filter((target): target is Address => target != null);
-  if (
-    !supportedTargets.some((supported) =>
-      isAddressEqual(params.authorized, supported),
-    )
-  ) {
-    throw new UnsupportedMidnightAuthorizationTargetError({
-      authorized: params.authorized,
-      chainId: params.chainId,
-      supportedTargets,
-    });
-  }
+  // Reject unknown operators before reading authorization state.
+  validateMidnightAuthorizationTarget({
+    chainId: params.chainId,
+    authorized: params.authorized,
+  });
   const midnight = getChainAddress(params.chainId, "midnight");
   const isAuthorized = await readContract(params.viemClient, {
     address: midnight,
